meta: Remove crystal / shards from global $PATH
This is a part of having separate flake.nix files for each project
and using devShells for the dependencies needed.

Note that using Crystal as a language seems less likely overall due to
the difficulty of building Crystal packages for nixpkgs and the lack of
tooling and library support compared to other languages like TypeScript
and Rust.