hyprland: Attempt to use hide on touch feature
This worked but broke when changing the display from e.g. 200% scaling
to 100% scaling. Since being able to easily change the scaling of
Hyprland was one of the main reasons I wanted to try Hyprland over
Phosh, I don't really have a use case for Hyprland on touch devices at
this moment, especially with all the drawbacks of using it on such a
small screen and without hardware acceleration.