ironbar: Add attempt at using volume module
This was my attempt at using the volume module, however I'd rather not
have to deal with using the small sliders that are extremely difficult
to see with the GTK theme I'm using.

I tried changing the colors with color, background, and border-color,
however I wasn't able to achieve the appearance I wanted, and disabling
the built-in popup results in a buggy left-clicking experience.