phone: Use maximum cpu frequency
Seems to be a bit more unstable but the performance difference is huge
enough that I'd rather run this than the minimum cpu frequency.

Battery life seems to be okay with around 3 hours of on-screen time and
low brightness. The massive increase in performance means that there's
much less time spent waiting for the phone, so the decrease in on-screen
time should be offset by how long you have to keep the phone on waiting
for it.