phone: Limit dirty_ratio / dirty_background_ratio
Been running this for a few days now and it seems to make the phone
significantly slower under load but reduces the probability of hanging
and crashing by quite a bit.

Plan to try this later with maximum CPU frequency to see if high CPU
frequency still crashes the PinePhone.