neovim: Drop remaining crystal support
As much as I love reading the Crystal programming language, it's clear
that there are more opportunities to be had with prioritizing Rust
instead. The ecosystem for Rust is vastly superior with higher quality
libraries and an LSP that's actually feature-complete, and I'd rather
deal with the known problems I'm aware of with Rust than the problems
I'm aware of with Crystal.

Rust won. Joking aside, the ecosystem for Rust is vastly superior, even
if the language is more difficult.