flake: Update stylix to fix startup issues with fish + tmux
Now it's possible to enjoy kitty image previews inside tmux, which is
something that zellij doesn't support.

Although tmux is slightly slower than zellij on startup, it has the
major advantage of a larger community with more time to iron out bugs
like the cursor jumping when scrolling in neovim with zellij.