phone: Update device UUIDs
Had to reinstall NixOS on the PinePhone again since a corrupted Nix
store broke everything and was unrecoverable due to not being able to
successfully repair specific files with the use of SSH substituters.

This time we will be trying the PinePhone without LUKS encryption to see
if this makes a difference in the performance of the device. Technically
encryption *isn't* supposed to make things slower in 2024 but the
PinePhone CPU is old enough that performance could've been affected.