flake: Switch to home-manager fork for librewolf support
I originally wanted to avoid maintaining my own forks of flake inputs to
simplify usage with the actual upstream if wanted, however the lack of
flakes supporting patches means that it's actually *easier* to maintain
my own repositories with the changes I want.

The main advantage of this is not having to wait for upstream. This also
means that I'm able to easily control which things I want to update and
when.