mobile-nixos: Limit audio to speaker/earpiece VoiceCall
This helps guarantee that phone call audio will work as long as the
phone wasn't suspended prior to the phone call, in which case the
pulseaudio server needs to be restarted first.

Note that this removes the possibility of using the PinePhone for tasks
where audio is particularly important, although it's likely more
enjoyable to accomplish those tasks on a more reliable device anyway.

I may come up with a solution that automates fixing the audio after
suspend in the future.