phone: Add script to record video
Note that we're using libopus and libx264 instead of flac and ffvhuff to
significantly reduce the file size and make it easier to play back the
recorded video on the PinePhone.

Note that there will be no preview while recording the video. The video
recording is also delayed a bit, so it's necessary to wait a few seconds
after you finish recording a video before pressing q to stop it.