Revert "overlays: Update picom to v12-rc2 for animations"
After trying the new picom animations for a bit they're a bit too janky
compared to the animations in Hyprland. The "geometry" trigger
replicated the tiling animation when swapping windows, however this came
at the cost of transparency being broken during the animation and the
easing in general not being comparable to Hyprland.

Further issues include the usage of animations breaking fade in/out
support for windows, which I found less intrusive than the appear /
disappear presets with not-so-subtle easing. Using the geometry trigger
also meant that switching tags in dwm had a slide in effect, which felt
disruptive compared to the usual feeling of dwm.

Other issues included the increased probability of things not being
drawn correctly (for example, an animation that "hangs" while in the
middle of animating) and a slightly annoying fade effect every time
windows changed in size such as when toggling fullscreen.

The config used is below for reference, however after trying picom
animations dwm feels considerably better without them.

```
animations = (
  {
    triggers = ["open"];
    preset = "appear";
    scale = 0.5;
  },
  {
    triggers = ["close"];
    preset = "disappear";
    scale = 0.5;
  },
  {
    triggers = ["geometry"];
    preset = "geometry-change";
    duration = 0.5;
  },
);
```