Revert "meta: Document usage of dwm on phone"
This was cool but there were a few issues such as me not being able to
find the infamous "dock patch" despite grepping for dock on
https://dwm.suckless.org/patches/, which resulted in the keyboard hiding
tiled windows and being below floating windows.

There was also an issue where starting corekeyboard from dwm would
result in styling not being applied compared to starting it from the
terminal.

A third issue was that from a quick attempt I couldn't get callaudiod to
start inside dwm, which was required for gnome-calls.

Overall it looks like I will be using Phosh until further notice on the
PinePhone, as I've already gotten it to work well enough for my use case.