meta: Replace mpdris2-rs with mpdris2
mpdris2 seems to have much greater feature support than mpdris2-rs,
notably the inclusion of lower notification urgency by default and the
ability to start mpdris2 without it showing a stopped notification.

The biggest feature, however, is the possibility to run multiple
mpdris2 instances at the same time. This was impossible to achieve
with mpdris2-rs since it would claim that the bus was already taken.

Notably this enables much smoother usage when running mpdris2 in both
Hyprland and dwm at the same time. Combined with dunst, this enables
seamless notifications between the two environments.