flake: remove mobile-nixos-vm
Previously we used a mobile-nixos config without the mobile-nixos
modules to start up a QEMU virtual machine on x86_64-linux without
having to worry about cross-compiling aarch64-linux.

Although this worked, the added complexity was not ideal and the
compilation time of the kernel with emulation was definitely not
ideal (5+ hours).

Note that I *may* add this again in the future, although I'd rather
set the system to x86_64-linux and cross-compile instead, which
mobile-nixos already supports but takes a large amount of ram.