desktop: add shotwell
This is by far the best software I found for my own personal image
management.

Although web interfaces like Immich and Szurubooru are cool and useful
when sharing a collection of images online, they pale in comparison to
the simplicity and low maintenance of using Shotwell, which also happens
to be considerably faster to browse large collections of images with.

By default images are stored in a YYYY/MM/DD hierarchy, which helps
prevent images from losing their creation date over time and enables
filtering images by their file name without having to add the date to
the file.

Furthermore, by organizing images in this way, the usage of tags is
encouraged, enabling images to belong to multiple groups of images
instead of a single folder categorization.